refactor: Move the styled_text module to lib and remove the styled_text module
